Refactor common memcache operations into a helper
